Inline the methods of ProcessorCount module.
